+++ OneBlink.SDK/models/GetSubmissionPdfRequest.cs
@@ -25,6 +25,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
-
+        public List<string> excludedCSSClasses
+        {
+            get; set;
+        }
     }
 }