+++ OneBlink.SDK/models/PDFConfiguration.cs
@@ -28,5 +28,9 @@ namespace OneBlink.SDK.Model
         {
             get; set;
         }
+        public List<string> excludedCSSClasses
+        {
+            get; set;
+        }
     }
 }